Fix observation lows/highs.

+++ tensortrade/environments/trading_environment.py
@@ -99,8 +99,8 @@ class TradingEnvironment(gym.Env, TimeIndexed):
 
         self._enable_logger = kwargs.get('enable_logger', False)
         self._observation_dtype = kwargs.get('dtype', np.float32)
-        self._observation_lows = kwargs.get('observation_lows', -np.iinfo(np.int32).max)
-        self._observation_highs = kwargs.get('observation_highs', np.iinfo(np.int32).max)
+        self._observation_lows = kwargs.get('observation_lows', -np.iinfo(np.int64).max)
+        self._observation_highs = kwargs.get('observation_highs', np.iinfo(np.int64).max)
         self._max_allowed_loss = kwargs.get('max_allowed_loss', 0.1)
 
         if self._enable_logger:
